h-signature implementations moved to a separate header.
